# apps/backend/src/modules/admin/admin-dashboard.controller.ts
import {
  Controller,
  Get,
  Query,
  UseGuards,
} from '@nestjs/common';
import { PrismaService } from '../prisma/prisma.service';
import { AdminGuard } from '../../common/guards/admin.guard';

@Controller('v1/admin')
@UseGuards(AdminGuard)
export class AdminDashboardController {
  constructor(private readonly prisma: PrismaService) {}

  @Get('dashboard/stats')
  async getDashboardStats() {
    const [
      totalPlants,
      verifiedPlants,
      totalOwners,
      totalConsumers,
      pendingVerifications,
      totalConversations,
      totalMessages,
      messagesLast24h,
      unreadMessages,
    ] = await Promise.all([
      this.prisma.plant.count(),
      this.prisma.plant.count({ where: { isVerified: true } }),
      this.prisma.owner.count(),
      this.prisma.consumer.count(),
      this.prisma.verificationRequest.count({ where: { status: 'PENDING' } }),
      this.prisma.conversation.count(),
      this.prisma.message.count(),
      this.prisma.message.count({
        where: {
          sentAt: { gte: new Date(Date.now() - 24 * 60 * 60 * 1000) },
        },
      }),
      this.prisma.message.count({ where: { readAt: null } }),
    ]);

    return {
      totalPlants,
      verifiedPlants,
      unverifiedPlants: totalPlants - verifiedPlants,
      totalOwners,
      totalConsumers,
      pendingVerifications,
      chat: {
        totalConversations,
        totalMessages,
        messagesLast24h,
        unreadMessages,
      },
    };
  }

  @Get('plants')
  async getPlants(
    @Query('page') page = '1',
    @Query('limit') limit = '20',
    @Query('search') search?: string,
    @Query('verified') verified?: string,
  ) {
    const pageNum = Math.max(1, parseInt(page, 10) || 1);
    const limitNum = Math.min(50, Math.max(1, parseInt(limit, 10) || 20));
    const skip = (pageNum - 1) * limitNum;

    const where: any = {};

    if (search) {
      where.OR = [
        { name: { contains: search, mode: 'insensitive' } },
        { address: { contains: search, mode: 'insensitive' } },
      ];
    }

    if (verified === 'true') {
      where.isVerified = true;
    } else if (verified === 'false') {
      where.isVerified = false;
    }

    const [plants, total] = await Promise.all([
      this.prisma.plant.findMany({
        where,
        include: {
          owner: {
            select: {
              id: true,
              name: true,
              email: true,
            },
          },
        },
        orderBy: { createdAt: 'desc' },
        skip,
        take: limitNum,
      }),
      this.prisma.plant.count({ where }),
    ]);

    return {
      plants,
      meta: {
        page: pageNum,
        limit: limitNum,
        total,
        totalPages: Math.ceil(total / limitNum),
      },
    };
  }

  @Get('owners')
  async getOwners(
    @Query('page') page = '1',
    @Query('limit') limit = '20',
    @Query('search') search?: string,
  ) {
    const pageNum = Math.max(1, parseInt(page, 10) || 1);
    const limitNum = Math.min(50, Math.max(1, parseInt(limit, 10) || 20));
    const skip = (pageNum - 1) * limitNum;

    const where: any = {};

    if (search) {
      where.OR = [
        { name: { contains: search, mode: 'insensitive' } },
        { email: { contains: search, mode: 'insensitive' } },
      ];
    }

    const [owners, total] = await Promise.all([
      this.prisma.owner.findMany({
        where,
        select: {
          id: true,
          name: true,
          email: true,
          phone: true,
          createdAt: true,
          _count: {
            select: {
              plants: true,
            },
          },
        },
        orderBy: { createdAt: 'desc' },
        skip,
        take: limitNum,
      }),
      this.prisma.owner.count({ where }),
    ]);

    return {
      owners,
      meta: {
        page: pageNum,
        limit: limitNum,
        total,
        totalPages: Math.ceil(total / limitNum),
      },
    };
  }

  @Get('consumers')
  async getConsumers(
    @Query('page') page = '1',
    @Query('limit') limit = '20',
    @Query('search') search?: string,
  ) {
    const pageNum = Math.max(1, parseInt(page, 10) || 1);
    const limitNum = Math.min(50, Math.max(1, parseInt(limit, 10) || 20));
    const skip = (pageNum - 1) * limitNum;

    const where: any = {};

    if (search) {
      where.OR = [
        { name: { contains: search, mode: 'insensitive' } },
        { email: { contains: search, mode: 'insensitive' } },
      ];
    }

    const [consumers, total] = await Promise.all([
      this.prisma.consumer.findMany({
        where,
        select: {
          id: true,
          name: true,
          email: true,
          phone: true,
          createdAt: true,
        },
        orderBy: { createdAt: 'desc' },
        skip,
        take: limitNum,
      }),
      this.prisma.consumer.count({ where }),
    ]);

    return {
      consumers,
      meta: {
        page: pageNum,
        limit: limitNum,
        total,
        totalPages: Math.ceil(total / limitNum),
      },
    };
  }
}