# specs/001-water-plant-discovery/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FlowGrid Water Plant Discovery API
  description: |
    RESTful API for FlowGrid Water Plant Discovery Platform V1.

    ## Authentication
    - Consumer App: JWT Bearer token (username/password login)
    - Owner App: JWT Bearer token (phone/password login)
    - Admin Dashboard: JWT Bearer token (username/password login)

    ## Rate Limiting
    - Public endpoints: 100 requests/minute per IP
    - Authenticated endpoints: 500 requests/minute per user

    ## Response Format
    All responses follow the standard envelope:
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "timestamp": "...", "requestId": "..." }
    }
    ```

    Error responses:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable message",
        "details": { ... }
      }
    }
    ```
  version: 1.0.0
  contact:
    name: FlowGrid API Support
    email: api@flowgrid.io
servers:
  - url: https://api.flowgrid.io/v1
    description: Production server
  - url: https://staging-api.flowgrid.io/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
...
paths:
...
  /plants:
    get:
      tags:
        - Plants (Public)
      summary: List plants near location
      description: |
        Get water plants within specified radius of a location.
        Results are sorted by distance by default.
      operationId: listPlants
      parameters:
        - name: lat
          in: query
          required: true
          description: Latitude of center point
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
            example: 17.385044
        - name: lng
          in: query
          required: true
          description: Longitude of center point
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
            example: 78.486671
        - name: radius
          in: query
          description: Search radius in meters (default 5000, max 50000)
          schema:
            type: integer
            default: 5000
            minimum: 100
            maximum: 50000
        - name: sort
          in: query
          description: Sort order for results
          schema:
            type: string
            enum: [distance, tds_asc, price_asc, price_desc]
            default: distance
        - name: filter_open
          in: query
          description: Filter to only open plants
          schema:
            type: boolean
            default: false
        - name: filter_verified
          in: query
          description: Filter to only verified plants
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of plants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /plants/{plantId}:
    get:
      tags:
        - Plants (Public)
      summary: Get plant details
      description: Get full details of a specific plant including photos and contact info
      operationId: getPlant
      parameters:
        - $ref: '#/components/parameters/PlantId'
      responses:
        '200':
          description: Plant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /owner/conversations:
    get:
      tags:
        - Owner Chat
      summary: List owner conversations
      description: Get all conversations for the authenticated owner, sorted by most recent activity
      operationId: listOwnerConversations
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Conversations list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /owner/conversations/{conversationId}/messages:
    get:
      tags:
        - Owner Chat
      summary: Get messages in conversation
      description: Get messages for a conversation, paginated, newest first
      operationId: getOwnerMessages
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: before
          in: query
          description: Get messages before this message ID (for pagination)
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: Messages list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Owner Chat
      summary: Send message
      description: Send a new message in the conversation
      operationId: sendOwnerMessage
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /owner/conversations/{conversationId}/read:
    post:
      tags:
        - Owner Chat
      summary: Mark messages as read
      description: Mark all messages in conversation as read up to a specific message
      operationId: markOwnerMessagesRead
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - upToMessageId
              properties:
                upToMessageId:
                  type: string
                  format: uuid
      responses:
        '204':
          description: Messages marked as read
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/plants:
    get:
      tags:
        - Admin Plants
      summary: List all plants
      description: Get list of all plants with search and filtering
      operationId: adminListPlants
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          description: Search by plant name
          schema:
            type: string
        - name: status
          in: query
          description: Filter by verification status
          schema:
            type: string
            enum: [unverified, pending, verified, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Plants list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPlantListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/plants/{plantId}:
    get:
      tags:
        - Admin Plants
      summary: Get plant details (admin)
      description: Get full plant details including owner information
      operationId: adminGetPlant
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PlantId'
      responses:
        '200':
          description: Plant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPlantDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # HEALTH CHECK
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  parameters:
    PlantId:
      name: plantId
      in: path
      required: true
      description: Plant unique identifier
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'UNAUTHORIZED'
              message: 'Invalid or expired token'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'FORBIDDEN'
              message: 'You do not have permission to perform this action'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'NOT_FOUND'
              message: 'Resource not found'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: 'VALIDATION_ERROR'
            message:
              type: string
              example: 'Validation failed'
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # ========== Auth ==========
...
    OwnerBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        email:
          type: string
...
    PlantListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        tdsReading:
          type: integer
          nullable: true
        pricePerLiter:
          type: number
          format: double
          nullable: true
        isOpen:
          type: boolean
        isVerified:
          type: boolean
        distance:
          type: number
          format: double
          description: Distance in meters from search location
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    PlantListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            plants:
              type: array
              items:
                $ref: '#/components/schemas/PlantListItem'
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - $ref: '#/components/schemas/PaginationMeta'
    PlantDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        operatingHours:
          type: string
          nullable: true
        tdsReading:
          type: integer
          nullable: true
        pricePerLiter:
          type: number
          format: double
          nullable: true
        description:
          type: string
          nullable: true
        photos:
          type: array
          items:
            type: string
            format: uri
        verificationStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        isVerified:
          type: boolean
        isOpen:
          type: boolean
        contact:
          type: object
          properties:
            phone:
              type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PlantDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/PlantDetail'
...
    AdminPlantListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        ownerName:
          type: string
        verificationStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        isOpen:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AdminPlantListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            plants:
              type: array
              items:
                $ref: '#/components/schemas/AdminPlantListItem'
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - $ref: '#/components/schemas/PaginationMeta'
    AdminPlantDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            plant:
              allOf:
                - $ref: '#/components/schemas/PlantDetail'
                - type: object
                  properties:
                    viewCount:
                      type: integer
            owner:
              allOf:
                - $ref: '#/components/schemas/OwnerBasic'
                - type: object
                  properties:
                    createdAt:
                      type: string
                      format: date-time

    # ========== Notifications ==========
...
    ConversationListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        plantId:
          type: string
          format: uuid
        plantName:
          type: string
        plantPhoto:
          type: string
          format: uri
          nullable: true
        otherParty:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            type:
              type: string
              enum: [consumer, owner]
        lastMessage:
          type: object
          nullable: true
          properties:
            content:
              type: string
              description: Truncated preview (first 50 chars)
            sentAt:
              type: string
              format: date-time
            isFromMe:
              type: boolean
        unreadCount:
          type: integer
        myRetention:
          type: string
          enum: [off, 24h, 7d, 30d]
        effectiveRetention:
          type: string
          enum: [off, 24h, 7d, 30d]
          description: Shorter of both party's settings
        createdAt:
          type: string
          format: date-time
    ConversationListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            conversations:
              type: array
              items:
                $ref: '#/components/schemas/ConversationListItem'
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - $ref: '#/components/schemas/PaginationMeta'
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        senderType:
          type: string
          enum: [consumer, owner]
        senderId:
          type: string
          format: uuid
        content:
          type: string
          description: Decrypted message content
        status:
          type: string
          enum: [sent, delivered, read]
        sentAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        readAt:
          type: string
          format: date-time
          nullable: true
        isFromMe:
          type: boolean
        expiresAt:
          type: string
          format: date-time
          nullable: true
    MessageListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            messages:
              type: array
              items:
                $ref: '#/components/schemas/ChatMessage'
            hasMore:
              type: boolean
              description: Whether there are more messages to load
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    SendMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
          description: Message text content
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            message:
              $ref: '#/components/schemas/ChatMessage'

    # ========== WebSocket Events (Documentation) ==========
    # Note: These are for documentation purposes. WebSocket uses ws://localhost:3000/chat
...